Add files via upload
Automatically debuffs Divine Protection. Just use "/bp on" in game to enable. "/bp off" to disable.

In 2012 I was the first Protection Paladin Raiding on Feenix to obtain Styleens and AQ40 Armament Shield and was about 3 Items away from BiS before the Server died. 
I had a Guide with over 500k Views and over 400 Comments at the time. 

The Meta has shifted very much be it on Classic and Turtle from those old Private Server times with higher Boss HP and Enemy Damage Output numbers. We had to gear Mitigation and wait for threat and this is prior to the Blessing of Kings abuse so my Guide is very dated, but for the sake of preservation I am going to include it here on an Addon that is not dated and very much relevant.

It is included in the files. Thank you.